trivial: restructure files relavent to sessions
