+++ aptos-move/aptos-vm/src/move_vm_ext/mod.rs
@@ -4,16 +4,14 @@
 //! MoveVM and Session wrapped, to make sure Aptos natives and extensions are always installed and
 //! taken care of after session finish.
 mod resolver;
-mod respawned_session;
-mod session;
+pub mod session;
 mod vm;
 mod warm_vm_cache;
 pub(crate) mod write_op_converter;
 
 pub use crate::move_vm_ext::{
     resolver::{AptosMoveResolver, AsExecutorView, AsResourceGroupView, ResourceGroupResolver},
-    respawned_session::RespawnedSession,
-    session::{SessionExt, SessionId},
+    session::SessionExt,
     vm::{get_max_binary_format_version, get_max_identifier_size, verifier_config, MoveVmExt},
 };
 use aptos_types::{access_path::AccessPath, state_store::state_key::StateKey};
@@ -21,6 +19,7 @@ use move_binary_format::errors::{PartialVMError, PartialVMResult};
 use move_core_types::{
     account_address::AccountAddress, language_storage::StructTag, vm_status::StatusCode,
 };
+pub use session::session_id::SessionId;
 
 pub(crate) fn resource_state_key(
     address: AccountAddress,
